Move AXI4 shim into the AxiWindow
Previously the AxiWindow only exposed a preWindow interface, and was supposed to be constructed by wrapping the postWindow interface.
Outside the module a AxiShimFF was created which provided the "postWindow" slave, requests for which get passed out through the shim's "master" interface.
It makes more semantic sense for the window to function like a shim with an extra port - so now it does.